+++ lib/eve_online.rb
@@ -63,6 +63,11 @@ require 'eve_online/esi/factions'
 require 'eve_online/esi/ancestries'
 require 'eve_online/esi/universe_types'
 require 'eve_online/esi/universe_type'
+require 'eve_online/esi/universe_regions'
+require 'eve_online/esi/universe_region'
+require 'eve_online/esi/universe_constellations'
+require 'eve_online/esi/universe_constellation'
+require 'eve_online/esi/universe_systems'
 
 require 'eve_online/esi/models/base'
 require 'eve_online/esi/models/loyalty_point'
@@ -106,6 +111,8 @@ require 'eve_online/esi/models/race'
 require 'eve_online/esi/models/bloodline'
 require 'eve_online/esi/models/faction'
 require 'eve_online/esi/models/ancestry'
+require 'eve_online/esi/models/region'
+require 'eve_online/esi/models/constellation'
 
 module EveOnline
 end