+++ lib/eve_online/esi/universe_types.rb
@@ -3,16 +3,28 @@
 module EveOnline
   module ESI
     class UniverseTypes < Base
-      API_ENDPOINT = 'https://esi.evetech.net/v1/universe/types/?datasource=%<datasource>s&page=1'
+      API_ENDPOINT = 'https://esi.evetech.net/v1/universe/types/?datasource=%<datasource>s&page=%<page>s'
+
+      attr_reader :page
+
+      def initialize(options = {})
+        super
+
+        @page = options.fetch(:page, 1)
+      end
 
       def universe_types_ids
         response
       end
 
+      def total_pages
+        resource.headers['x-pages']&.to_i
+      end
+
       def scope; end
 
       def url
-        format(API_ENDPOINT, datasource: datasource)
+        format(API_ENDPOINT, datasource: datasource, page: page)
       end
     end
   end